♻️ Refactor publish and other into advanced settings
