+++ likecoin/js/admin-settings/src/pages/MainSettingLayout.js
@@ -5,7 +5,6 @@ function MainSettingLayout() {
   return <><nav>
       <Link to="">{__('General', 'likecoin')}</Link>
       <Link to="advanced">{__('Advanced', 'likecoin')}</Link>
-      <Link to="other">{__('Other', 'likecoin')}</Link>
       <Link to="about">{__('About', 'likecoin')}</Link>
     </nav>
     <Outlet />
